# docs/conexao-hicode/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HII Motor API",
    "version": "1.2.0",
    "description": "API single-user para o backend Hicode. Extensao de observabilidade v1 independente da ponte legada. O motor e a autoridade do estado. Partida do daemon somente por POST administrativo com opt-in explicito. Nao faz merge."
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
    "/v1/diagnostico": {
      "get": {
        "operationId": "diagnosticarMotor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Resposta do motor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Diagnostico v1 sem inferencia, somente API administrativa. Checks com estado, escopo, duracao e horario; timeout global de 30 segundos."
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Revisao para If-Match quando aplicavel"
              }
            }
          },
          "default": {
            "description": "Erro de autenticacao, validacao, revisao, conflito ou operacao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recuperacoes/previa": {
      "post": {
        "operationId": "previaRecuperacao",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PacoteRecuperacao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta do motor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Revisao para If-Match quando aplicavel"
              }
            }
          },
          "default": {
            "description": "Erro de autenticacao, validacao, revisao, conflito ou operacao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recuperacoes/importar": {
      "post": {
        "operationId": "importarRecuperacao",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{8,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "pacote": {
                    "$ref": "#/components/schemas/PacoteRecuperacao"
                  },
                  "hash": {
                    "type": "string"
                  }
                },
                "required": [
                  "pacote",
                  "hash"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta do motor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Revisao para If-Match quando aplicavel"
              }
            }
          },
          "default": {
            "description": "Erro de autenticacao, validacao, revisao, conflito ou operacao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tarefas/{id}/recuperacao": {
      "get": {
        "operationId": "diagnosticarRecuperacao",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{3,12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta do motor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Revisao para If-Match quando aplicavel"
              }
            }
          },
          "default": {
            "description": "Erro de autenticacao, validacao, revisao, conflito ou operacao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tarefas/{id}/restaurar-configuracao": {
      "post": {
        "operationId": "restaurarConfiguracao",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{8,128}$"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{3,12}$"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "hash": {
                    "type": "string"
                  }
                },
                "required": [
                  "hash"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta do motor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Revisao para If-Match quando aplicavel"
              }
            }
          },
          "default": {
            "description": "Erro de autenticacao, validacao, revisao, conflito ou operacao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/motor/iniciar": {
      "post": {
        "operationId": "iniciarMotor",
        "description": "Opt-in HII_API_AUTOSTART=1 e API administrativa. Sem argumentos de processo. Serializa e limita tentativas; nao retoma cards pausados.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {},
                "required": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta do motor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Revisao para If-Match quando aplicavel"
              }
            }
          },
          "default": {
            "description": "Erro de autenticacao, validacao, revisao, conflito ou operacao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/v1/motor/status": {
      "get": {
        "operationId": "estadoMotor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Resposta do motor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "protocolo": {
                      "const": 1
                    },
                    "estado": {
                      "enum": [
                        "ligado",
                        "desligado",
                        "degradado",
                        "desconhecido"
                      ]
                    },
                    "versao": {
                      "type": "string"
                    },
                    "versaoEmExecucao": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "fila": {
                      "type": "string"
                    },
                    "consultadoEm": {
                      "type": "string"
                    },
                    "motivo": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "protocolo",
                    "estado",
                    "versao",
                    "versaoEmExecucao",
                    "fila",
                    "consultadoEm",
                    "motivo"
                  ]
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Revisao para If-Match quando aplicavel"
              }
            }
          },
          "default": {
            "description": "Erro de autenticacao, validacao, revisao, conflito ou operacao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Erro": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "erro": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "codigo": {
                "type": "string"
              },
              "mensagem": {
                "type": "string"
              }
            },
            "required": [
              "codigo",
              "mensagem"
            ]
          }
        },
        "required": [
          "erro"
        ]
      },
...
    }
  }
}